+++ src/StateForge.Replication/StateForgeClusterMemberRole.cs
@@ -1,0 +1,8 @@
+namespace StateForge.Replication
+{
+    public enum StateForgeClusterMemberRole
+    {
+        Primary = 0,
+        Replica = 1
+    }
+}